feat: move svd_stack and fft_solver out of utils

+++ include/utils/progress_bar.hpp
@@ -33,7 +33,7 @@ namespace progresscpp {
 
             std::cout << "[";
 
-            for (int i = 0; i < bar_width; ++i) {
+            for (int i = 0; i < (int)bar_width; ++i) {
                 if (i < pos) std::cout << complete_char;
                 else if (i == pos) std::cout << ">";
                 else std::cout << incomplete_char;
